% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include.R
\name{rspiro_check_input}
\alias{rspiro_check_input}
\title{Check input data for errors}
\usage{
rspiro_check_input(spiro_val, somat_val)
}
\arguments{
\item{spiro_val}{List of spirometry measurements}

\item{somat_val}{List of somatometric measurements}
}
\value{
Invisible. Stops if it finds an error.
}
\description{
This function checks the supplied input for errors
}
\keyword{internal}
