% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscores_JRS.R
\name{zscore_JRS}
\alias{zscore_JRS}
\title{Convert spirometric values to z-scores using JRS (2014) equations}
\usage{
zscore_JRS(
  age,
  height,
  gender = 1,
  FEV1 = NULL,
  FVC = NULL,
  VC = NULL,
  FEV1FVC = NULL
)
}
\arguments{
\item{age}{Age in years}

\item{height}{Height in meters}

\item{gender}{Gender (1 = male, 2 = female) or a factor with two levels (first = male). Default is 1.}

\item{FEV1}{Forced Expiratory Volume in 1 second (lt)}

\item{FVC}{Forced Vital Capacity (lt)}

\item{VC}{Vital Capacity (lt)}

\item{FEV1FVC}{FEV1 / FVC}
}
\value{
If only one spirometry argument is supplied, the function
returns a numeric vector. If more are supplied, the function returns
a data.frame with the same number of columns.
}
\description{
This function takes absolute spirometry measurements (FEV1, FVC, VC, FEV1/FVC)
in lt plus demographic data (age, height and gender) and converts
them to z-scores based on the JRS (Japanese Respiratory Society 2014) equations.
}
\details{
At least one of the spirometric measurement arguments must be set (i.e. be
non-\code{NULL}). Arguments \code{age}, \code{height}, and \code{gender}
must be vectors of length equal to the length of the
spirometric measurement vector(s), or of length one, in which case their
value is recycled. If any input vector is not of equal length, the function
stops with an error.
}
\examples{
# Random data, 4 patients, one parameter supplied (FEV1)
zscore_JRS(age=seq(25,40,4), height=c(1.8, 1.9, 1.75, 1.85),
      gender=c(2,1,2,1), FEV1=c(3.5, 4, 3.6, 3.9))

}
